New operator.Pipeline API
